==> janice/1_IntroToCompBoot.r <==

print("This is a code cell")

# Do some arithmetic here using the R kernel. Insert another R cell below and try some more.

# Make the next cell into a markdown cell (using the keyboard shortcut) and type some text. 
# Can you make it bold? Italicized? 
# Can you figure out how to make headings?
