==> janice/8_Pilot2019_getstarcounts.ipynb <==
### Load paths and libraries

library(foreach)
library(fs)
library(tidyverse)

data_root_dir <- "/data/hts_2019_data/"
raw_fastq_dir <- paste(data_root_dir, "hts2019_pilot_rawdata", sep = "")
metadata_file <- file.path(raw_fastq_dir, "2019_pilot_metadata.tsv")
count_dir <- paste(data_root_dir, "hts2019_pilot_counts", sep = "")

# Create directories

Below is an illustration of our folder structure.
```
scratch
└── bioinf_intro
└── analysis_output
    ├── out -> the folder to store all our output data files
    └── img -> the folder to store all our images  
```

curdir <- "/home/jovyan/work/scratch/analysis_output"
outdir <- file.path(curdir, "out")
imgdir <- file.path(curdir, "img")

system(paste("mkdir -p", outdir), intern = TRUE)
system(paste("mkdir -p", imgdir), intern = TRUE)

count_suffix <- "_ReadsPerGene.out.tab"
bam_suffix <- "_Aligned.sortedByCoord.out.bam"

### Load STAR Count Data

list.files(count_dir, pattern = paste0(count_suffix,"$"), full.names = FALSE) %>%
      path_filter(regex="_M1") -> countfiles
countfiles

length(countfiles)

So there are 96 files about 2019 pilot data under this folder. Each file is generated from a fastq file using STAR.

mycombine <- function(df1, df2) {
    # Combine two data frames by gene names
    #
    # Args:
    #   df1 (Dataframe): the first count data
    #   df2 (Dataframe): the second count data
    #
    # Returns:
    #   (Dataframe) The combined data frame of df1 and df2
    full_join(df1, df2, by = "gene")
}


mystarfile <- function(rootdir, stardir) {
    # Get the absolute paths of a file
    #
    # Args:
    #   rootdir  (Character): the directory of the folder
    #   stardir (Character): the filename
    #
    # Returns:
    #   (Character) the directory of the input file
    file.path(rootdir, stardir)
}

# Data type for each column
coltypes <- list(col_character(), col_integer(), col_integer(), col_integer())

read the count files and combine them

out <- foreach(file = countfiles, .combine = mycombine) %do% {
    cntfile <- mystarfile(count_dir, file)
    readr::read_tsv(cntfile, col_names = FALSE, col_types = coltypes ) %>%
        dplyr::select(X1, X4) %>%
            dplyr::rename_(.dots=setNames(names(.), c("gene", file)))
}


dim(out)

out[1:6, 1:6]

### Gather and spread the genes to get a count matrix (genecounts)

out %>%
    slice(-(1:4)) %>%
    gather(expid, value, -gene) %>% 
    spread(gene, value) -> genecounts

genecounts[1:4, 1:6]

### Gather and spread the first four rows to nmisc

For nmisc, we will take the first 4 rows of out since those are the summarizing features. Next, we want to transform the data frame so that it is in statistical format (the samples are the rows and the feature types are the columns). Using a combination of gather and spread, we can transpose our matrix into the desired format.

out %>%
    slice(1:4) %>%
    gather(expid, value, -gene) %>% 
    spread(gene, value) %>%
    rename_(.dots=setNames(names(.), c("expid", "namb", "nmulti", "nnofeat","nunmap"))) -> nmisc

nmisc[1:6, ]

### Gather and spread the gene rows

For each samples, we want to sum up all the counts, so we can create a variable denoting the number of total genes mapped for each sample by summing across the rows.

out %>%
    slice(-(1:4)) %>%
    gather(expid, value, -gene) %>% 
    spread(gene, value) %>%
    mutate(ngenemap=rowSums(.[-1])) %>%
    select(expid, ngenemap) -> ngene

ngene[1:6, ]

### merge in the 4 misc counts and add summaries

So far, we can create a comprehensive data frame mapresults which will combine ngene with nmisc. This data frame will have summarizing mapping features in addition to proportion features.

ngene %>%
    full_join(nmisc, by="expid") %>%
    mutate(depth = as.integer(ngenemap + namb + nmulti + nnofeat + nunmap)) %>%
    mutate(prob.gene = ngenemap / depth) %>%
    mutate(prob.nofeat = nnofeat / depth) %>%
    mutate(prob.unique = (ngenemap+nnofeat) / depth) -> mapresults

mapresults[1:4, ]

outfile <- file.path(outdir, "hts-pilot-2019.RData")
save(mapresults, genecounts, file=outfile)
tools::md5sum(outfile)
