==> janice/10_Pilot2018_counts.ipynb <==
library(tidyverse)
library(foreach)
library(ggplot2)
library(RColorBrewer)

curdir <- "/home/jovyan/work/scratch/analysis_output"
outdir <- file.path(curdir, "out")
imgdir <- file.path(curdir, "img")

### Import Counts

sum_cntfile <- file.path(outdir, "hts-pilot-2019.RData")
attach(sum_cntfile)
tools::md5sum(sum_cntfile)

dim(genecounts)

dim(mapresults)

### Import metadata file

metadtfile <- "/data/hts_2019_data/hts2019_pilot_rawdata/2019_pilot_metadata.tsv"
tools::md5sum(metadtfile)
mtdf<-readr::read_tsv(metadtfile)
dim(mtdf)

head(mtdf)

str(mtdf)

### Check to see if we can match the labels of counts
### to those in the metadata file

myregex <- "_[A-Z](100|[1-9][0-9]?)_L00[1-4]_ReadsPerGene.out.tab"

mtdf$Label %>%
    setdiff(gsub(myregex, "", mapresults$expid))

### Add the "Label" to the count matrix and mapping results,
### and the merge in phenotype data (by Label)

### Add "Label" to genecounts

genecounts %>%
    mutate(Label=gsub(myregex, "", c(expid))) ->
    annogenecnts

dim(genecounts)

dim(annogenecnts)

### Collapse the gene counts within each label

annogenecnts %>%
    select(-expid) %>%
    group_by(Label) %>%
    summarize_all(sum) %>%
    gather(gene, value, -Label) %>% 
    spread(Label, value) ->
    annogenecnts0

dim(annogenecnts0)

annogenecnts0[1:6, 1:6]

### Add "Label" to read map results and merge in phenotype data

mapresults %>%
    mutate(Label=str_replace(expid, myregex, "")) %>%
    full_join(mtdf, by="Label") ->
    annomapres

dim(annomapres)

grpvars<-vars(Label, genotype, condition, sample_year,
              libprep_person,enrichment_method)
sumvars<-vars(prob.gene,prob.nofeat,prob.unique,depth)

annomapres %>%
  group_by_at(grpvars) %>%
  summarize_at(sumvars,mean) -> annomapres0

outfile <- file.path(outdir, "HTS-Pilot-Annotated-STAR-counts.RData")
save(mtdf, annogenecnts0, annomapres0, annogenecnts, annomapres, file = outfile)
tools::md5sum(outfile)

### Figures for mapping results

mygeom <-  geom_point(position = position_jitter(w = 0.3, h = 0))
mypal <- scale_colour_manual(name="",  values =brewer.pal(3,"Set1"))
mytheme <-  theme(axis.text.x = element_text(angle = 90, hjust = 1)) + theme_bw()
myfacet <- facet_grid(genotype ~ group, drop=TRUE, scales="free_x", space="free")

p1 <- ggplot(annomapres, aes(x=factor(Label), y=prob.unique, shape=genotype, color=condition))+
        myfacet+mygeom+mytheme+mypal

png(file.path(imgdir, "p1.png"), height=480*2, width=480*2)
plot(p1)
graphics.off()

p2 <- ggplot(annomapres, aes(x=Label, y=prob.gene, shape=genotype, color=condition))+
        myfacet+mygeom+mytheme+mypal
  
png(file.path(imgdir, "p2.png"), height=480*2, width=480*2)
plot(p2)
graphics.off()

p3 <- ggplot(annomapres, aes(x=Label, y=prob.nofeat, shape=genotype, color=condition))+
        myfacet+mygeom+mytheme+mypal
png(file.path(imgdir, "p3.png"), height=480*2, width=480*2)
plot(p3)
graphics.off()

p4 <- ggplot(annomapres, aes(x=Label, y=depth, shape=genotype, color=condition))+
        myfacet+mygeom+mytheme+mypal
png(file.path(imgdir, "p4.png"), height=480*2, width=480*2)
plot(p4)
graphics.off()
