==> janice/10_Rlibraries_and_bioconductor.ipynb <==
# R libraries and Bioconductor

## Packages and Libraries

R is at heart a collection of 'packages'. There is a 'base' system that contains the truly basic commands, such as the assignment operator `->` or the command to create a vector. In addition to that, there are 'standard R' packages that are included when you install the R kernel (in the Jupyter notebook), or 'R' as a program to run either at the command line or with Rstudio. (I've shown some examples of these different ways to run R in class).

### Libraries

Many packages, even those included in [standard R] (https://www.r-project.org/), will need to be 'loaded' to be used. In other words, they exist on your computer (or in your container), but the R kernel doesn't know about them. This is because if it did, R would be using computer memory (RAM) to remember all their functions and variables. If all the available packages were loaded, you might not have any RAM left!

A consequence of this is that you often have to tell R explicitly that you want to use a particular package. You do that using `library`. Let's read in the titanic data set to have something to play with.



titanic <- read.csv("titanic.csv")

head(titanic)

There is a cool R function that will allow us to look at some random rows from a data frame. It's called `sample_n`. Let's try it:

sample_n(titanic, 10)

Oops. It turns out `sample_n` is in the dplyr package. It's installed in your container - but R doesn't know that! Let's tell R we want to use it:

library(dplyr)

sample_n(titanic, 10)

### Installed and installing packages

Now, `dplyr` is actually not part of standard R. It's *installed* separately. There are a multitude of R packages out there. Anyone can write one (yes, even you!!!). They are shared with the public using the [CRAN archive.] (https://cran.r-project.org/) In order to be listed in CRAN, packages need to meet specific criteria for documentation purposes, testing, etc.

You can check to see what packages are installed using `installed.packages()`

installed.packages()

You can install new packages using the command `install.packages()`


install.packages("auk", lib = "/home/jovyan/work")

remove.packages("auk", lib = "/home/jovyan/work")

## BioConductor

CRAN is home to many, many R packages. But there is a whole other world out there when it comes to bioinformatics in R. It's called [BioConductor](https://bioconductor.org/). BioConductor is a comprehensive toolkit for all things having to do with high-throughput sequencing data processing and analysis. In this course, we will use the BioConductor package `DESeq2` to perform differential expression analysis. It's the end of the pipeline, after QC, clipping and trimming, aligning and counting. 

### Installing BioConductor packages

BioConductor has it's own installation procedure (and it's own criteria for documentation, testing, etc.) - separate from CRAN. Let's have a look at the page for [DESeq2](https://bioconductor.org/packages/release/bioc/html/DESeq2.html)

### S4 Objects

R has two main types of 'objects' that can work with 'methods' (functions that are specific to those objects): S3 and S4. S3 objects are really just lists with the `class` attribute set. For example:

cat <- list(name = "Dash", age = 10)
class(cat) <- "Pet"

dog <- list(name = "Tibbs", age = 8)
class(dog) <- "Pet"

print.Pet <- function(x){
    print(paste("Name:", x$name, "Age:", x$age))
}

print(cat)

S4 objects are a bit more stringent. We won't contruct our own right now, just mention that they are created with constructor functions that make sure there is consistency. 

Bioconductor makes extensive use of S4 objects. One of the main features of these is inheritance. This means that if an object is a 'child' of an existing object, the child inherits all the features and functions of the parent.

Let's look at some of the simpler Bioconductor objects that will be used to build up more complicated structures.

suppressPackageStartupMessages({
  library(Biostrings)
})
dna <- DNAStringSet( c("AAACTG", "CCCAACCA") ) #DNAStringSet is a constructor function for the class DNAStringSet

dna
reverseComplement(dna)  #This is a method that works on a DNAStringSet object





# This doesn't work

dna2 <- c("AAACTG", "CCCAACCA")
reverseComplement(dna2)

Where to find help:
  
- `class(dna)`
- `methods(class = "DNAStringSet")`
- `?"DNAStringSet"`, `?"reverseComplement,DNAStringSet-method"` (tab
                                                                 completion!)
- `browseVignettes(package="Biostrings")`
- https://bioconductor.org/packages
- https://support.bioconductor.org

[Biostrings]: https://bioconductor.org/packages/Biostrings




methods(class = "DNAStringSet")

#### GenomicRanges

suppressPackageStartupMessages({
  library(GenomicRanges)
})
gr <- GRanges(
  c("chr1:10-19", "chr1:15-24", "chr1:30-39")
)
gr

- Closed-interval (start and end coordinates include in range, like
                   Ensembl; UCSC uses 1/2-open)
- 1-based (like Ensembl; UCSC uses 0-based)

Operations

help("GRanges")
width(gr)

shift(gr, 1)
shift(gr, c(1, 2, 3))


`GRangesList`

- List of `GRanges`, e.g., exons

gene <- c("A", "A", "B")
grl <- splitAsList(gr, gene)
grl

## Airway example

We'll walk through an example using a sample data set called 'airway'. Airway is an object of type 'SummarizedExperiment'. This kind of object is the basis for many objects used in Bioconductor packages.

library("airway")
data("airway")
se <- airway

str(se)

[This tutorial](https://bioconductor.org/packages/devel/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html) gives a great introduction to the SummarizedExperiment object. We'll take a peek, and then move on to DESeq2 

assays(se)

assays(se)$counts

rowRanges(se)

colData(se)

metadata(se)

# Just a list - we can add elements

metadata(se)$formula <- counts ~ dex + albut

metadata(se)

# subset the first five transcripts and first three samples
se[1:5, 1:3]

assays(se[1:5,1:3])$counts


library("DESeq2")


dds <- DESeqDataSet(se, design = ~ cell + dex)
dds



# remove rows with less than 10 total transcripts

keep <- rowSums(counts(dds)) >= 10
dds <- dds[keep,]

dds$dex


# Specify reference level

dds$dex <- factor(dds$dex, levels = c("untrt","trt"))

#alternative
dds$dex <- relevel(dds$dex, ref = "untrt")

dds$dex

ddsDE <- DESeq(dds)
res <- results(ddsDE)
res
